+++ misk-web/web/packages/@misk/cli/lib/commands/ci-build.d.ts
@@ -1,0 +1,3 @@
+export declare const command = "ci-build";
+export declare const desc = "run a fast, clean, webpack production build";
+export declare function handler(): Promise<void>;