+++ misk-web/web/packages/@misk/cli/lib/commands/clean.d.ts
@@ -1,0 +1,3 @@
+export declare const command = "clean";
+export declare const desc = "remove build directory and other temporary files";
+export declare function handler(): Promise<void>;