+++ misk-web/web/packages/@misk/cli/lib/commands/index.d.ts
@@ -1,0 +1,2 @@
+export { handler as build } from "./build";
+export { handler as prebuild } from "./prebuild";