+++ misk-web/web/packages/@misk/cli/lib/commands/prebuild.d.ts
@@ -1,0 +1,3 @@
+export declare const command = "prebuild";
+export declare const desc = "consume miskTab.json and write necessary build files";
+export declare function handler(): Promise<void>;