+++ misk-web/web/packages/@misk/cli/lib/commands/start.d.ts
@@ -1,0 +1,3 @@
+export declare const command = "start";
+export declare const desc = "start webpack development server";
+export declare function handler(): Promise<void>;