+++ misk-web/web/packages/@misk/cli/lib/commands/zip.d.ts
@@ -1,0 +1,3 @@
+export declare const command = "zip";
+export declare const desc = "zip source code for tab";
+export declare function handler(): Promise<void>;