+++ misk-web/web/packages/@misk/cli/lib/utils/migrate.d.ts
@@ -1,0 +1,1 @@
+export declare const migrate: () => Promise<void>;