+++ misk-web/web/packages/@misk/cli/lib/utils/templates/prettier.js
@@ -1,0 +1,16 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.prettier = {
+    prettier: {
+        arrowParens: "avoid",
+        bracketSpacing: true,
+        jsxBracketSameLine: false,
+        printWidth: 80,
+        semi: false,
+        singleQuote: false,
+        tabWidth: 2,
+        trailingComma: "none",
+        useTabs: false
+    }
+};
+//# sourceMappingURL=prettier.js.map