+++ misk-web/web/packages/@misk/cli/lib/utils/templates/tsconfig.d.ts
@@ -1,0 +1,6 @@
+import { IMiskTabJSON } from "../index";
+export declare const createTsconfig: (miskTab: IMiskTabJSON) => {
+    compileOnSave: boolean;
+    compilerOptions: any;
+    include: string[];
+};