+++ misk-web/web/packages/@misk/dev/CHANGELOG.md
@@ -1,6 +1,14 @@
 # Change Log - @misk/dev
 
-This log was last generated on Thu, 07 Feb 2019 19:11:09 GMT and should not be manually modified.
+This log was last generated on Thu, 21 Feb 2019 19:28:11 GMT and should not be manually modified.
+
+## 0.1.4
+
+Thu, 21 Feb 2019 19:28:11 GMT
+
+### Updates
+
+- Add @misk/simplenetwork to base webpack config
 
 ## 0.1.3